milahaina: gps: Fixes ubsan shift-out-of-bounds SIGABRT
There are now global sanitizers that will
be applied to all modules, so we no longer
need to decalre local sanitizer options.
One of the local sanitizer options caused a
ubsan shift-out-of-bounds SIGABRT in several
location modules in code that does not appear
to have anything problematic with it.

CRs-fixed: 2898374
Change-Id: Icb7e3b2cf35be7c0a945eff16cea2b3a4f30e4a8
Signed-off-by: Mohd Faraz <androiabledroid@gmail.com>